+++ packages/types/src/appearance.ts
@@ -11,6 +11,7 @@ import type {
   SelectId,
   UserPreviewId,
 } from './elementIds';
+import type { EnterpriseProvider } from './enterpriseAccount';
 import type { OAuthProvider } from './oauth';
 import type { SamlIdpSlug } from './saml';
 import type { BuiltInColors, TransparentColor } from './theme';
@@ -172,7 +173,10 @@ export type ElementsConfig = {
   socialButtonsProviderIcon: WithOptions<OAuthProvider | Web3Provider, LoadingState>;
   socialButtonsProviderInitialIcon: WithOptions<OAuthProvider | Web3Provider, LoadingState>;
 
-  enterpriseButtonsProviderIcon: WithOptions<SamlIdpSlug, LoadingState>;
+  enterpriseButtonsProviderIcon: WithOptions<EnterpriseProvider, LoadingState>;
+
+  providerIcon: WithOptions<OAuthProvider | Web3Provider | SamlIdpSlug, LoadingState>;
+  providerInitialIcon: WithOptions<OAuthProvider | Web3Provider | SamlIdpSlug, LoadingState>;
 
   alternativeMethods: WithOptions;
   alternativeMethodsBlockButton: WithOptions<OAuthProvider | Web3Provider, LoadingState>;
@@ -397,7 +401,6 @@ export type ElementsConfig = {
   badge: WithOptions<'primary' | 'actionRequired'>;
   notificationBadge: WithOptions;
   buttonArrowIcon: WithOptions;
-  providerIcon: WithOptions<OAuthProvider | Web3Provider>;
   spinner: WithOptions;
 };