+++ packages/expo/CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 2.3.3
+
+### Patch Changes
+
+- Updated dependencies [[`1c0b5001f7f975a2f3f54ad692526ecf7257847e`](https://github.com/clerk/javascript/commit/1c0b5001f7f975a2f3f54ad692526ecf7257847e), [`6217a3f7c94311d49f873214bd406961e0b8d6f7`](https://github.com/clerk/javascript/commit/6217a3f7c94311d49f873214bd406961e0b8d6f7), [`e1d715683a057e4f6166095e551861c4b35b7ac6`](https://github.com/clerk/javascript/commit/e1d715683a057e4f6166095e551861c4b35b7ac6), [`1783025cdb94c447028704c2121fa0b8af785904`](https://github.com/clerk/javascript/commit/1783025cdb94c447028704c2121fa0b8af785904)]:
+  - @clerk/clerk-js@5.34.0
+  - @clerk/types@4.33.0
+  - @clerk/shared@2.13.0
+  - @clerk/clerk-react@5.15.4
+
+## 2.3.2
+
+### Patch Changes
+
+- Updated dependencies [[`152019b07e75a31e354e8ea50a07afb907ebf320`](https://github.com/clerk/javascript/commit/152019b07e75a31e354e8ea50a07afb907ebf320), [`ff4ebeba6c2a77c247a946070b56bdb2153d1588`](https://github.com/clerk/javascript/commit/ff4ebeba6c2a77c247a946070b56bdb2153d1588)]:
+  - @clerk/clerk-js@5.33.1
+
 ## 2.3.1
 
 ### Patch Changes